Main

+++ scripts/setup-clean.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 #set -x
-source "/vagrant/scripts/common.sh"
+if [ -d /vagrant/scripts ];then
+    source "/vagrant/scripts/common.sh"
+fi
 
 # 删除安装目录
-rm -rf ${INSTALL_PATH}/azkaban-3.84.4
+[ -d ${INSTALL_PATH}/azkaban-3.84.4 ] && rm -rf ${INSTALL_PATH}/azkaban-3.84.4
+[ -f /root/anaconda-ks.cfg ] && rm -rf /root/anaconda-ks.cfg  
+[ -f /root/original-ks.cfg ] && rm -rf /root/original-ks.cfg
